+++ composer.json
@@ -14,5 +14,8 @@
     "require": {
         "php": "^8.1",
         "psr/http-message": "^1.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9.5"
     }
 }
