+++ src/Domain/Protocols/TaskRepositoryInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Anglesson\Exemplo\Domain\Protocols;
+
+use Anglesson\Exemplo\Domain\Entity\Task;
+
+interface TaskRepositoryInterface
+{
+    // /** @return Task[] */
+    // public function list(): array;
+    // public function findById(string $id): ?Task;
+    public function save(Task $tarefa): Task;
+    // public function update(Task $tarefa): Task;
+    // public function delete(Task $tarefa): void;
+}
